Game+Tests: Fix an issue when matching repeating letters in a guess word
Before that change, `Game` would match repeating letters multiple times against secret. In case of full/partial match this would lead to confusion as player would think that this letter would appear multiple times in a secret.

For example, secret word could be 'AFTER' and the guess word 'AWARE'. Before this change, game would match first 'A' letter as full match and the second one as partial match.

When doing word matching, we will now masking indexes in a secret word that were used. This means for example above after first 'A' match, secret would be '_FTER' and so we wouldn't match 'A' second time. This also means that if there are multiple of same letters in a secret and guess, they would be matched correctly as we're masking letters using indexes.