UI: Implement `FlipWidget`
At the moment its API is a little bit awkward, but it works reasonably well. Creating `AnimationController` outside of the `FlipWidget` is required if we actually want to control when to start an animation -- which is generally how it works. However, the awkward part comes from the the fact the only thing we need to initialize is the duration. Any other modification most likely will produce unexpected results.

For example, setting lowerBound and upperBound will have an effect on how many rotation widget will do in a span of duration. Ideally, of course, widget should do only one rotation and no more.

Creation of `AnimationController` also forces its creation to have `TickerProviderStateMixin` mixin.