UI: Implement `LetterTile` flip on word submission
It uses newly implemented `FlipWidget` to display two different tiles: one tile is simply a letter in a box and the other one is filled with color.

`LetterTile` constructor now accepts row/column instead of a `Letter`. This allows the widget itself to understand whether it should flip at all and when.